requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
strands-agents
boto3
